requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dateutil==2.8.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6
typing_extensions==4.8.0
openai==1.3.5
numpy==1.26.2
matplotlib==3.8.2 